requirements.txt
discord.py[voice]==1.2.3
feedparser==5.2.1
rasa==1.3.7
requests==2.22.0
